更新增加华为快引用的配置文件的处理
